+++ ICD.Profound.ConnectPRO/Themes/UserInterface/IViews/Common/ICameraControlView.cs
@@ -58,6 +58,11 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.IViews.Common
 		/// </summary>
 		event EventHandler<UShortEventArgs> OnCameraButtonPressed;
 
+		/// <summary>
+		/// Raised when a tab button is pressed.
+		/// </summary>
+		event EventHandler<UShortEventArgs> OnTabButtonPressed;
+
 		#endregion
 
 		/// <summary>
@@ -85,5 +90,18 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.IViews.Common
 		/// <param name="index"></param>
 		/// <param name="selected"></param>
 		void SetCameraSelected(ushort index, bool selected);
+		
+		/// <summary>
+		/// Sets the selected state of a tab button.
+		/// </summary>
+		/// <param name="index"></param>
+		/// <param name="selected"></param>
+		void SetTabSelected(ushort index, bool selected);
+
+		/// <summary>
+		/// Sets the visibility of the tab button
+		/// </summary>
+		/// <param name="visible"></param>
+		void SetTabVisibility(bool visible);
 	}
 }